requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
deep-translator==1.11.4
distro==1.9.0
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
google-generativeai==0.8.3
google-cloud-texttospeech==2.16.3
PyAudio==0.2.14
pydantic==2.6.4
pydantic_core==2.16.3
pygame==2.5.2
requests==2.31.0
googlemaps==4.10.0
sniffio==1.3.1
soupsieve==2.5
SpeechRecognition==3.10.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
werkzeug==3.0.1
